fix: implement --version argument and calculate latest xnat-tools and validator versions based on numbering
The "version" argument was previously allowed but not implemented; passing it to run_xnat2bids.py succeeded, but then it would be passed as an argument in the xnat2bids call and cause a crash because it isn't an allowed argument. Now, if that argument is passed in the config file, it is used to choose the correct singularity image version and then deleted from the parameter list before it is passed to xnat2bids. If the --version argument is not used, the most recent version of xnat-tools is used based on tags (not based on the date the image was pulled to Oscar, which may or may not be the most recent).

BIDS validator previously used the "latest" tag, which we are no longer pulling to Oscar. This fix uses the same function to choose the most recent numbered tag.