feat: add new script anonymize_dicoms.py
This script copies a directory of DICOMs, making modifications to the DICOM headers as specified in a configuration .csv file.

For example, this demo .csv:
Tag,Value
PatientBirthDate,CLEAR
PerformingPhysicianName,DELETE
PatientID,Anonymous

tells the script to clear the "PatientBirthDate" field, delete the "PerformingPhysicianName" field, and set the "PatientID" field to a value of "Anonymous".