fix: handle spaces in xnat2bids arg vals
If a value passed to xnat-tools contained a space (e.g. the sequence name anat-t1w_acq-memprage RMS_ND), it was not being parsed correctly. Wrapping it in quotes fixes the problem, but in order to do that I also had to change the syntax for the sbatch command, because shlex.split was not respecting the quotes and still splitting on the space.