fix: only add flag to xnat-tools call if true in toml file
Previously, the value in the toml file was ignored, so setting overwrite=false or validate-frames=false would pass --overwrite and --validate_frames to xnat-tools. Now if the parameter is not included in the toml file, or is set to false, it will not be passed as a flag to xnat-tools.