fix: error if both project and XNAT sessions are present in toml file
If both are present, the sessions list gets ignored and all sessions for the project are converted, which is unlikely to be the desired behavior if sessions are specified. Instead just error:

Both project and sessions are defined in configuration file.
Please specify with either (project & subject) OR session (XNAT Accession #)